Update BMICalculator.java
im not making any changes right now, I am just leaving a comment. I tried to used gitbash to upload the src file but it wouldnt let me, it kept giving me a bunch of errors, so i decided to upload the file like this so it could at least be there.